refactor to `json_schemas`
