+++ lib/src/protocol/json_schemas/balance_schema.dart
@@ -8,7 +8,7 @@ class BalanceSchema {
     "additionalProperties": false,
     "currencyCode": {
       "type": "string",
-      "description": "ISO 3166 currency code string"
+      "description": "ISO 4217 currency code string"
     },
     "available": {
       "$ref": "definitions.json#/definitions/decimalString",
