+++ lib/src/protocol/json_schemas/close_schema.dart
