+++ lib/src/protocol/json_schemas/definitions_schema.dart
