+++ lib/src/protocol/json_schemas/message_schema.dart
