+++ lib/src/protocol/json_schemas/reputation_schema.dart
