+++ lib/src/protocol/json_schemas/resource_schema.dart
