+++ lib/src/protocol/json_schemas/rfq_private_schema.dart
