+++ lib/src/protocol/json_schemas/rfq_schema.dart
